feat: add a log module for controlling the logger
